requirements.txt
# Core ML/RL libraries
torch>=2.0.0
torchvision>=0.15.0
gymnasium>=0.29.0
stable-baselines3>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data processing and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web framework for interactive interface
flask>=2.3.0
flask-cors>=4.0.0

# Database and data handling
requests>=2.31.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0 